Improve the evalRegex syntax.
The previous syntax didn't make much sense, or at least was inconsistent with kantan.xpath. The new syntax is slightly less powerful though: evalRegex only accepts regular expressions as strings or Regex, where it used to accept any A that had a Compiler instance. I see it as an acceptable trade of for consistency, especially since using Pattern or scala.matching.Regex should be discouraged.