Lays the groundwork for a proper generic module.
The current implementation is lacking, in that it has working decoders for coproducts and hlists, but doesn't go the extra mile of turning these into decoder instances for ADTs and case classes. I realise that this might not be feasible in a satisfactory fashion because of https://github.com/milessabin/export-hook/issues/24.